Accesspoint/requirements.txt
bleak==0.20.2
PyYAML==6.0
requests==2.28.2
